requirements.txt
htmlmin
beautifulsoup4
PyYAML
webassets
slugify
jinja2


